Оновлено кінцеві точки контактів

+++ src/main/java/ua/natusvincere/echat/contact/ContactController.java
@@ -3,13 +3,14 @@ package ua.natusvincere.echat.contact;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.w3c.dom.stylesheets.LinkStyle;
+import ua.natusvincere.echat.exception.ForbiddenException;
 
 import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/contacts")
+@RequestMapping("/api/v1/contacts")
 public class ContactController {
 
     private final ContactService contactService;
@@ -23,11 +24,17 @@ public class ContactController {
             @RequestBody AddContactRequest request,
             Principal principal
     ) {
+        if (principal == null) {
+            throw new ForbiddenException("You are not authorized");
+        }
         return ResponseEntity.ok(contactService.addContact(request, principal));
     }
 
     @GetMapping
     public ResponseEntity<List<ContactResponse>> findAll(Principal principal) {
+        if (principal == null) {
+            throw new ForbiddenException("You are not authorized");
+        }
         return ResponseEntity.ok(contactService.findAll(principal));
     }
 
@@ -36,6 +43,9 @@ public class ContactController {
             @PathVariable("id") UUID id,
             Principal principal
     ) {
+        if (principal == null) {
+            throw new ForbiddenException("You are not authorized");
+        }
         contactService.deleteContact(id, principal);
         return ResponseEntity.noContent().build();
     }
